create migration successfully,still create mysql container

+++ src/main/resources/db/migration/V1__creation_des_tables.sql
@@ -1,12 +1,12 @@
 CREATE TABLE client (
-                        id INT AUTO_INCREMENT PRIMARY KEY,
+                        id BIGINT AUTO_INCREMENT PRIMARY KEY,
                         nom VARCHAR(100) NOT NULL,
                         email VARCHAR(150),
                         phone VARCHAR(30)
 );
 
 CREATE TABLE chauffeur (
-                           id int AUTO_INCREMENT PRIMARY KEY,
+                           id BIGINT AUTO_INCREMENT PRIMARY KEY,
                            nom VARCHAR(100) NOT NULL,
                            phone VARCHAR(30),
                            permis_type VARCHAR(50),
@@ -14,7 +14,7 @@ CREATE TABLE chauffeur (
 );
 
 CREATE TABLE vehicule (
-                          id int AUTO_INCREMENT PRIMARY KEY,
+                          id BIGINT AUTO_INCREMENT PRIMARY KEY,
                           matricule VARCHAR(50) UNIQUE NOT NULL,
                           capacite INT,
                           disponible BOOLEAN ,
@@ -23,7 +23,7 @@ CREATE TABLE vehicule (
 
 
 CREATE TABLE livraison (
-                           id int AUTO_INCREMENT PRIMARY KEY,
+                           id BIGINT AUTO_INCREMENT PRIMARY KEY,
                            date_livraison DATE,
                            adresse_depart VARCHAR(255),
                            adresse_destination VARCHAR(255),
